Implement takes_context for simple_tag
Add a `PyContext` wrapper struct to allow passing `Context` for use from
Python. This uses an `Arc` to make it unnecessary to implement `Clone`
for `Context`, which would be difficult given the nested `Py` types in
`Context.request` and `Context.context`.

In `SimpleTag::render`, use `std::mem::replace` to convert the
`&mut context` into an owned type that can be wrapped as a `PyContext`.

After calling the Python tag render implementation, use `std::mem::swap`
to convert the `PyContext.context` back into `&mut context` for
continued use by other tag render methods. This currently relies on the
`Arc` having only one reference, but unfortunately this isn't
guaranteed. A badly behaved tag can take a reference to the `PyContext`
that lasts after it returns (for example using the `global` keyword).